+++ client/Login_old.css
@@ -1,19 +1,3 @@
-body {
-  background: #F2CCEF !important;
-}
-.tutor-name {
-  font-size: 24pt;
-  font-weight: bold;
-  color: #2f4971;
-  margin-top: 1rem;
-}
-.tutor-description {
-  font-size: 16pt;
-  color: #4b5563;
-}
-.text-container {
-  text-align: center;
-}
 .control-button {
   margin-top: 1rem;
   padding: 0.5rem 1rem;
