+++ client/entry-client.jsx
@@ -7,6 +7,11 @@ import "./base.css";
 import Register from "./pages/register";
 
 const container = document.getElementById("root");
+
+// note, to add new endpoint. you need to edit entry-server.jsx and entry-client.jsx
+// if you want the path to be public, go edit server.js at line 39
+// const publicPaths = ["/login", "/register", "/newendpoint"];
+
 createRoot(container).render(
   <StrictMode>
     <BrowserRouter>
