+++ client/entry-server.jsx
@@ -6,6 +6,10 @@ import App from "./components/App";
 import Login from "./pages/login";
 import Register from "./pages/register";
 
+// note, to add new endpoint. you need to edit entry-server.jsx and entry-client.jsx
+// if you want the path to be public, go edit server.js at line 39
+// const publicPaths = ["/login", "/register", "/newendpoint"];
+
 export function render(url) {
   const html = renderToString(
     <StrictMode>
